Feat: 공통 API 호출 함수를 분리하여 작성
