+++ src/app/(providers)/(root)/detail/(movenew)/allreviews/[id]/page.tsx
@@ -2,29 +2,15 @@
 
 import React from 'react';
 import { useRouter, useParams } from 'next/navigation';
-import axios from 'axios';
 import { useQuery } from '@tanstack/react-query';
+import { fetchReviews, Review } from '@/utils/api/detail/reviews';
 import { averageRatings } from '@/utils/averageRating';
 import ReactStars from 'react-rating-stars-component';
 
-interface Review {
-  id: string;
-  created_at: string;
-  user_id: string;
-  post_id: string;
-  content: string;
-  rating: number;
-}
-
 const AllReviewsPage = () => {
-  const { id: postId } = useParams();
+  const params = useParams();
+  const postId = Array.isArray(params.id) ? params.id[0] : params.id;
   const router = useRouter();
-
-  // 리뷰 데이터를 가져오는 함수
-  const fetchReviews = async (): Promise<Review[]> => {
-    const response = await axios.get(`/api/detail/reviews/${postId}`);
-    return response.data;
-  };
 
   // React Query를 사용하여 리뷰 데이터 가져오기
   const {
@@ -33,7 +19,7 @@ const AllReviewsPage = () => {
     error
   } = useQuery<Review[]>({
     queryKey: ['reviews', postId],
-    queryFn: fetchReviews,
+    queryFn: () => fetchReviews(postId),
     enabled: !!postId // postId가 있을 때만 쿼리 실행
   });
